requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
exceptiongroup==1.3.0
groq==0.33.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.3
pydantic_core==2.41.4
pyTelegramBotAPI==4.29.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
telebot==0.0.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Pillow==10.4.0
transformers==4.57.1
sentence-transformers==5.1.2
